third_party/WebKit/LayoutTests/external/wpt/css/motion-1/parsing/offset-path-parsing-valid-expected.txt

This is a testharness.js-based test.
PASS e.style['offset-path'] = "none" should set the property value
PASS Serialization should round-trip after setting e.style['offset-path'] = "none"
PASS e.style['offset-path'] = "ray(0rad closest-side)" should set the property value
PASS Serialization should round-trip after setting e.style['offset-path'] = "ray(0rad closest-side)"
PASS e.style['offset-path'] = "ray(0.25turn closest-corner contain)" should set the property value
PASS Serialization should round-trip after setting e.style['offset-path'] = "ray(0.25turn closest-corner contain)"
PASS e.style['offset-path'] = "ray(200grad farthest-side)" should set the property value
PASS Serialization should round-trip after setting e.style['offset-path'] = "ray(200grad farthest-side)"
PASS e.style['offset-path'] = "ray(270deg farthest-corner contain)" should set the property value
PASS Serialization should round-trip after setting e.style['offset-path'] = "ray(270deg farthest-corner contain)"
PASS e.style['offset-path'] = "ray(-720deg sides)" should set the property value
PASS Serialization should round-trip after setting e.style['offset-path'] = "ray(-720deg sides)"
PASS e.style['offset-path'] = "ray(calc(180deg - 45deg) farthest-side)" should set the property value
PASS Serialization should round-trip after setting e.style['offset-path'] = "ray(calc(180deg - 45deg) farthest-side)"
PASS e.style['offset-path'] = "path('m 0 0 h -100')" should set the property value
PASS Serialization should round-trip after setting e.style['offset-path'] = "path('m 0 0 h -100')"
PASS e.style['offset-path'] = "path('M 0 0 L 100 100 M 100 200 L 200 200 Z L 300 300 Z')" should set the property value
PASS Serialization should round-trip after setting e.style['offset-path'] = "path('M 0 0 L 100 100 M 100 200 L 200 200 Z L 300 300 Z')"
FAIL e.style['offset-path'] = "url(\"http://www.example.com/index.html#polyline1\")" should set the property valueassert_not_equals: got disallowed value ""
FAIL Serialization should round-trip after setting e.style['offset-path'] = "url(\"http://www.example.com/index.html#polyline1\")"assert_equals: expected "url(\"http://www.example.com/index.html#polyline1\")" but got ""
FAIL e.style['offset-path'] = "circle(100px)" should set the property valueassert_not_equals: got disallowed value ""
FAIL Serialization should round-trip after setting e.style['offset-path'] = "circle(100px)"assert_equals: expected "circle(100px)" but got ""
FAIL e.style['offset-path'] = "margin-box" should set the property valueassert_not_equals: got disallowed value ""
FAIL Serialization should round-trip after setting e.style['offset-path'] = "margin-box"assert_equals: expected "margin-box" but got ""
FAIL e.style['offset-path'] = "inset(10% 20% 30% 40%) border-box" should set the property valueassert_not_equals: got disallowed value ""
FAIL Serialization should round-trip after setting e.style['offset-path'] = "inset(10% 20% 30% 40%) border-box"assert_equals: expected "inset(10% 20% 30% 40%) border-box" but got ""
FAIL e.style['offset-path'] = "fill-box ellipse(50% 60%)" should set the property valueassert_not_equals: got disallowed value ""
FAIL Serialization should round-trip after setting e.style['offset-path'] = "fill-box ellipse(50% 60%)"assert_equals: expected "ellipse(50% 60%) fill-box" but got ""
Harness: the test ran to completion.
